tests/wpt/tests/dom/nodes/Document-implementation.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Document.implementation</title>
<link rel=help href="https://dom.spec.whatwg.org/#dom-document-implementation">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  var implementation = document.implementation;
  assert_true(implementation instanceof DOMImplementation,
              "implementation should implement DOMImplementation");
  assert_equals(document.implementation, implementation);
}, "Getting implementation off the same document");

test(function() {
  var doc = document.implementation.createHTMLDocument();
  assert_not_equals(document.implementation, doc.implementation);
}, "Getting implementation off different documents");
</script>
